+++ workspaces/arborist/lib/edge.js
@@ -4,6 +4,7 @@
 const util = require('util')
 const npa = require('npm-package-arg')
 const depValid = require('./dep-valid.js')
+const OverrideSet = require('./override-set.js')
 
 class ArboristEdge {
   constructor (edge) {
@@ -267,7 +268,7 @@ class Edge {
         this.#error = 'PEER LOCAL'
       } else if (!this.satisfiedBy(this.#to)) {
         this.#error = 'INVALID'
-      } else if (this.overrides && this.#to.edgesOut.size && this.#to.constructor.doOverrideSetsConflict(this.overrides, this.#to.overrides)) {
+      } else if (this.overrides && this.#to.edgesOut.size && OverrideSet.doOverrideSetsConflict(this.overrides, this.#to.overrides)) {
         // Any inconsistency between the edge's override set and the target's override set is potentially problematic.
         // But we only say the edge is in error if the override sets are plainly conflicting.
         // Note that if the target doesn't have any dependencies of their own, then this inconsistency is irrelevant.
