+++ workspaces/arborist/lib/node.js
@@ -1386,24 +1386,6 @@ class Node {
     }
   }
 
-  static findSpecificOverrideSet (first, second) {
-    for (let overrideSet = second; overrideSet; overrideSet = overrideSet.parent) {
-      if (overrideSet.isEqual(first)) {
-        return second
-      }
-    }
-    for (let overrideSet = first; overrideSet; overrideSet = overrideSet.parent) {
-      if (overrideSet.isEqual(second)) {
-        return first
-      }
-    }
-    log.silly('Conflicting override sets', first, second)
-  }
-
-  static doOverrideSetsConflict (first, second) {
-    return (this.findSpecificOverrideSet(first, second) === undefined)
-  }
-
   updateOverridesEdgeInRemoved (otherOverrideSet) {
     // If this edge's overrides isn't equal to this node's overrides, then removing it won't change newOverrideSet later.
     if (!this.overrides || !this.overrides.isEqual(otherOverrideSet)) {
@@ -1412,7 +1394,7 @@ class Node {
     let newOverrideSet
     for (const edge of this.edgesIn) {
       if (newOverrideSet && edge.overrides) {
-        newOverrideSet = Node.findSpecificOverrideSet(edge.overrides, newOverrideSet)
+        newOverrideSet = OverrideSet.findSpecificOverrideSet(edge.overrides, newOverrideSet)
       } else {
         newOverrideSet = edge.overrides
       }
@@ -1451,7 +1433,7 @@ class Node {
     if (this.overrides.isEqual(otherOverrideSet)) {
       return false
     }
-    const newOverrideSet = Node.findSpecificOverrideSet(this.overrides, otherOverrideSet)
+    const newOverrideSet = OverrideSet.findSpecificOverrideSet(this.overrides, otherOverrideSet)
     if (newOverrideSet) {
       if (!this.overrides.isEqual(newOverrideSet)) {
         this.overrides = newOverrideSet
